Fix reverb signed multiplication bug and clear SDL audio queue on sound stop
Agent-Logs-Url: https://github.com/awe444/sa2/sessions/eb43aaf5-80b5-4ca8-961d-d6671c0aaef2

+++ sa1/src/platform/shared/audio/m4a_sound_mixer.c
@@ -93,7 +93,11 @@ static void SampleMixer(struct SoundMixerState *mixer, u32 scanlineLimit, u16 sa
         u16 i = 0;
         do {
             fixed8_24 s = tmp1[0] + tmp1[1] + tmp2[0] + tmp2[1];
-            s = (s * reverb) >> 9;
+            // Cast to s64 to prevent overflow and preserve sign for negative values.
+            // Without this, s32 * u32 promotes to unsigned, and the logical right shift
+            // produces wrong positive values for negative sums (~50% of audio samples),
+            // causing noise and distortion.
+            s = (s32)(((s64)s * reverb) >> 9);
             tmp1[0] = tmp1[1] = s;
             tmp1 += 2;
             tmp2 += 2;
